Create coordinate_calculator.py
simple way to figure out where the source and destination coords are of the arm. Assumptions made is that all holes, tray borders, and distances between the holes are 5 but obviously that can be changed. Also assumptions made that the destination tray is placed directly to the right of the source tray